requirements.txt
pandas~=1.4.1
pre-commit==2.20.0
pymysql~=1.0.2
python-dotenv~=0.19.2
sqlalchemy~=1.4.27